feat(cs):modified files controllers and dtos

+++ Proyecto-REMI-WebApi/Proyecto-REMI-WebApi/Models/DTO's/PedidoDto.cs
@@ -22,6 +22,7 @@ namespace Proyecto_REMI_WebApi.Models.DTO_s
 
         public List<pedidoDetalleDto> detallesP { get; set; } = new();
 
-        public double valorPedido { get; set; }
+        public decimal valorPedido { get; set; }
+
     }
 }
